+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   - Added API to show and save survey questions one at a time
   - Added API to resume a survey
   - Sheets can now be partially validated to support incremental sheet updates
+- **Check Changes**
+  - Reduced the number of checks that need to be run when updating sheets
 - **Survey Changes**
   - Removed excess padding for section descriptions on public surveys
 - **Variable Changes**